+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # kdam Changelog (DD/MM/YYYY)
 
+## 0.1.8 (dev)
+
+Faetures:
+
+- Optimizations for printing mechanism.
+
+Changes:
+
+- `kdam::Bar.output` field is replaced with `kdam::Bar.writer`.
+- `kdam::term::Colorizer` trait is now only implemented for `&str` and `String`.
+- `kdam::term::init` function is removed and merged insided `kdam::term::colour` function.
+- `kdam::term::get_columns` is now public.
+- `kdam::BarProgress` trait is replaced with `kdam::TqdmIterator`.
+
 ## 0.1.7 (18/05/2022)
 
 Bug Fixes: